Add URL Visit Tool
Added a tool that returns the parsed and cleaned contents of a webpage when a URL is input. The agent can use this to find out more information about the program. Also reorganized a bunch of stuff in the repo like making a prompts folder and a utils folder which houses the parser.